Refactor: Remove code duplication between stablecoin.jsx and reservecoin.jsx (#110)
